Add packages

+++ src/communication/incoming/room/G_USRS.h
@@ -31,7 +31,7 @@ void G_USRS(player *p, incoming_message *message) {
     outgoing_message *players;
     players = om_create(28); // "@\"
 
-    for (int i = 0; i < list_size(room->users); i++) {
+    for (size_t i = 0; i < list_size(room->users); i++) {
         player *room_player;
         list_get_at(room->users, i, (void*)&room_player);
         append_user_list(players, room_player);
